web_app/backend/requirements.txt
flask
torch
torchvision
Pillow
numpy
scikit-image
huggingface_hub
